[Refactory] watch 패키지 이전 및 생성
수강신청 기능을 구현하기 위해서
기존 enrollment 패키지에 있던 영상 수강 페이지를 watch 패키지를 생성해 하위에 이동했습니다.
