+++ src/main/java/com/wanted/legendkim/domain/enrollment/EnrollmentController.java
@@ -1,28 +1,4 @@
 package com.wanted.legendkim.domain.enrollment;
 
-import com.wanted.legendkim.domain.enrollment.dto.WatchInfoResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
+// 수강 신청 관련 API 엔드포인트는 여기에 추가 예정
 
-@Controller
-@RequiredArgsConstructor
-@RequestMapping("/user/enrollments")
-public class EnrollmentController {
-
-    private final EnrollmentService enrollmentService;
-
-    @GetMapping("/{enrollmentId}/watch")
-    public String watch(
-            @PathVariable Long enrollmentId,
-            Model model) {
-
-        WatchInfoResponse response = enrollmentService.getWatchInfo(enrollmentId);
-        model.addAttribute("watchInfo", response);
-        return "movie/movie";
-    }
-
-}
