+++ src/main/java/com/wanted/legendkim/domain/section/Section.java
@@ -3,7 +3,7 @@ package com.wanted.legendkim.domain.section;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.wanted.legendkim.domain.course.Course;
-import com.wanted.legendkim.domain.enrollment.dto.SectionSummary;
+import com.wanted.legendkim.domain.watch.dto.SectionSummary;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
